# src/main/java/docs/openapi-gen.yaml
{"openapi":"3.1.0",
 "info":
   {"title":"OpenAPI definition"
     ,"version":"v0"},
 "servers":[{"url":"http://localhost:8081","description":"Generated server url"}],
 "paths":
   {"/users":
     {"get":
        {"tags":["local-user-controller"],
         "operationId":"getAllLocalUsers",
         "responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},
                      "200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},
      "put":{"tags":["local-user-controller"],"operationId":"updateLocalUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocalUser"}}},"required":true},
             "responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},
                          "200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LocalUser"}}}}}},
      "post":{"tags":["local-user-controller"],"operationId":"createLocalUser","requestBody":{"description":"New user example","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequestDto"},
                                                                                                                                                              "examples":{"Hans":{"summary":"User example 1","description":"Hans","value":{"name":"Hans","surname":"Müller","email":"hm@e.de","password":"12345","postAddress":"Berliner Str. 55, 10115 Berlin, Germany"}},"Émilie":{"summary":"User example 2","description":"Émilie","value":{"name":"Émilie","surname":"Dubois","email":"emi@e.fr","password":"12345","postAddress":"12 Rue de Rivoli, 75001 Paris, France"}}}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}}}},"/tickets":{"get":{"tags":["ticket-controller"],"operationId":"getAllTickets","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Ticket"}}}}}}},"post":{"tags":["ticket-controller"],"operationId":"createTicket","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Ticket"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"202":{"description":"Accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Ticket"}}}}}}},"/login":{"post":{"tags":["authentification-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/users/{id}":{"get":{"tags":["local-user-controller"],"operationId":"getLocalUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LocalUser"}}}}}},"delete":{"tags":["local-user-controller"],"operationId":"deleteLocalUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}}},"/tickets/{id}":{"get":{"tags":["ticket-controller"],"operationId":"getTicketById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Ticket"}}}}}}}},"components":{"schemas":{"LocalUser":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"surname":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"postAddress":{"type":"string"},"role":{"type":"string","enum":["ROLE_USER","ROLE_ADMINISTRATOR"]},"tickets":{"type":"array","items":{"$ref":"#/components/schemas/Ticket"},"uniqueItems":true}}},"Ticket":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"price":{"type":"number","format":"double"},"localUser":{"$ref":"#/components/schemas/LocalUser"}}},"CreateUserRequestDto":{"type":"object","properties":{"name":{"type":"string"},"surname":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"postAddress":{"type":"string"},"tickets":{"type":"array","items":{"$ref":"#/components/schemas/Ticket"},"uniqueItems":true}}},"UserResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nane":{"type":"string"},"surname":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"token":{"type":"string"}}}}}}